Set user data on submit
